Refactor useProvider, move provider related constants to provider.ts
