+++ packages/buidler-core/test/internal/buidler-evm/helpers/constants.ts
@@ -1,5 +1,5 @@
 // reused from ethers.js
-import { BN, bufferToHex, privateToAddress, toBuffer } from "ethereumjs-util";
+import { BN, toBuffer } from "ethereumjs-util";
 
 export const INFURA_URL = `https://mainnet.infura.io/v3/84842078b09946638c03157f83405213`;
 export const INFURA_NETWORK_ID = 1;
@@ -38,30 +38,3 @@ export const LAST_TX_HASH_OF_10496585 = toBuffer(
 );
 
 export const NULL_BYTES_32 = toBuffer(`0x${"0".repeat(64)}`);
-
-export const DEFAULT_HARDFORK = "istanbul";
-export const DEFAULT_NETWORK_NAME = "TestNet";
-export const DEFAULT_CHAIN_ID = 123;
-export const DEFAULT_NETWORK_ID = 234;
-export const DEFAULT_BLOCK_GAS_LIMIT = 6000000;
-export const DEFAULT_USE_JSON_RPC = false;
-export const DEFAULT_ALLOW_UNLIMITED_CONTRACT_SIZE = false;
-
-export const DEFAULT_ACCOUNTS = [
-  {
-    privateKey:
-      "0xe331b6d69882b4cb4ea581d88e0b604039a3de5967688d3dcffdd2270c0fd109",
-    balance: new BN(10).pow(new BN(18)),
-  },
-  {
-    privateKey:
-      "0xe331b6d69882b4cb4ea581d88e0b604039a3de5967688d3dcffdd2270c0fd10a",
-    balance: new BN(10).pow(new BN(18)),
-  },
-];
-export const DEFAULT_ACCOUNTS_ADDRESSES = DEFAULT_ACCOUNTS.map((account) =>
-  bufferToHex(privateToAddress(toBuffer(account.privateKey))).toLowerCase()
-);
-export const DEFAULT_ACCOUNTS_BALANCES = DEFAULT_ACCOUNTS.map(
-  (account) => account.balance
-);
