+++ packages/buidler-core/test/internal/buidler-evm/provider/modules/web3.ts
@@ -3,7 +3,7 @@ import { keccak256, toBuffer } from "ethereumjs-util";
 
 import { bufferToRpcData } from "../../../../../src/internal/buidler-evm/provider/output";
 import { setCWD } from "../../helpers/cwd";
-import { PROVIDERS } from "../../helpers/useProvider";
+import { PROVIDERS } from "../../helpers/providers";
 
 describe("Web3 module", function () {
   PROVIDERS.forEach((provider) => {
